rename firebase to realtime-db
